// dashboard/types/api.ts
export interface AutomodConfig {
  guild_id: number;
  enabled: boolean;
  punishments: Record<string, string>;
  ignored_roles: number[];
  ignored_channels: number[];
  logging_channel: number | null;
  blacklist_words?: string[];
}